added styling to header and footer and moved nav hamburger to footer
