+++ README.md
@@ -85,12 +85,6 @@ Imported from example to header component
 
 
 
-
-
-
-
-
-
 # Getting Started with Create React App
 
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
