+++ public/index.html
@@ -3,7 +3,11 @@
 <html lang="en">
 
   <head>
+    
+    <link rel="stylesheet" href="./style.css"></link>
+
     <title>Team Bengal</title>
+
     <meta charset="utf-8" />
 
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
