+++ public/style.css
@@ -1,0 +1,5 @@
+.Header {
+    background-color: black;
+    color: aqua;
+    text-align: center;
+}
