+++ src/components/Footer.js
@@ -1,8 +1,54 @@
-import React from 'react'
+import React, { useState } from 'react'
 
-const Footer = () => {
+import {
+  Collapse,
+  Navbar,
+  NavbarToggler,
+  NavbarBrand,
+  Nav,
+  NavItem,
+  NavLink,
+  UncontrolledDropdown,
+  DropdownToggle,
+  DropdownMenu,
+  DropdownItem,
+  NavbarText,
+} from 'reactstrap';
+
+const Footer = (args) => {
+
+  const [isOpen, setIsOpen] = useState(true);
+  const toggle = () => setIsOpen(!isOpen);
+
+
   return (
-    <div>Footer</div>
+    <>
+      <img src='./components/resources/Logo_cat.png' />
+
+      <Navbar {...args} color="info" fixed='bottom'>
+
+      <NavbarBrand href="/">Team Bengal</NavbarBrand>
+      <NavbarToggler onClick={toggle} />
+
+      <Collapse isOpen={isOpen} navbar>
+
+        <Nav navbar>
+          <NavItem><NavLink href="/catindex">See the cats</NavLink></NavItem>
+
+          <NavItem><NavLink href="/catshow">See a cat</NavLink></NavItem>
+
+          <NavItem><NavLink href="/catnew">Create a new cat</NavLink></NavItem>
+
+          <NavItem><NavLink href="/catedit">Edit a feline</NavLink></NavItem>
+
+          <NavItem><NavLink href="/*">woopsies</NavLink></NavItem>
+
+        </Nav>
+
+      </Collapse>
+
+    </Navbar>
+    </>
   )
 }
