+++ src/components/Header.js
@@ -2,9 +2,7 @@ import React from 'react'
 
 const Header = () => {
   return (
-    <>
-      <img src='./components/resources/Logo_cat.png' />
-    </>
+    <div className='Header'>Where cats come to meet</div>
   )
 }
